adjust likes, dislikes, commentNum when member deleted

+++ src/main/java/com/innovation/stockstock/post/dto/PostResponseDto.java
@@ -16,8 +16,8 @@ public class PostResponseDto {
     private String stockName;
     private boolean isDoneLike;
     private boolean isDoneDisLike;
-    private Long likes;
-    private Long dislikes;
+    private int likes;
+    private int dislikes;
     private int commentNum;
     private Member member;
     private List<CommentResponseDto> comments;
